# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Comment management service for Minisource",
        "title": "Comment Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5010",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/admin/comments/bulk-moderate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Bulk moderate comments",
                "parameters": [
                    {
                        "description": "Bulk moderation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.BulkModerateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.BulkModerateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/comments/pending": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get pending comments for moderation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_minisource_comment_internal_models.Comment"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/comments/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Permanently delete a comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/comments/{id}/moderate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Moderate a comment (approve/reject)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Moderation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.ModerateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.Comment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/comments/{id}/pin": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Pin or unpin a comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Pin data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.PinCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.Comment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_minisource_comment_internal_models.Attachment": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mimeType": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "type": {
                    "description": "image, video, file",
                    "type": "string"
                },
                "uploadedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_minisource_comment_internal_models.Comment": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_minisource_comment_internal_models.Attachment"
                    }
                },
                "authorAvatar": {
                    "type": "string"
                },
                "authorEmail": {
                    "type": "string"
                },
                "authorId": {
                    "description": "Author info",
                    "type": "string"
                },
                "authorName": {
                    "type": "string"
                },
                "content": {
                    "description": "Content",
                    "type": "string"
                },
                "contentHtml": {
                    "description": "Sanitized HTML",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Timestamps",
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "deletedBy": {
                    "type": "string"
                },
                "depth": {
                    "description": "Depth for nested replies",
                    "type": "integer"
                },
                "dislikeCount": {
                    "type": "integer"
                },
                "editHistory": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_minisource_comment_internal_models.EditRecord"
                    }
                },
                "flaggedWords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "isAnonymous": {
                    "type": "boolean"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "isEdited": {
                    "type": "boolean"
                },
                "isPinned": {
                    "description": "Features",
                    "type": "boolean"
                },
                "likeCount": {
                    "type": "integer"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "moderatedAt": {
                    "type": "string"
                },
                "moderatedBy": {
                    "type": "string"
                },
                "parentId": {
                    "description": "For replies",
                    "type": "string"
                },
                "pinnedAt": {
                    "type": "string"
                },
                "pinnedBy": {
                    "type": "string"
                },
                "reactionCounts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "rejectionReason": {
                    "type": "string"
                },
                "replyCount": {
                    "description": "Stats",
                    "type": "integer"
                },
                "reportCount": {
                    "type": "integer"
                },
                "resourceId": {
                    "description": "ID of the resource being commented on",
                    "type": "string"
                },
                "resourceType": {
                    "description": "e.g., \"product\", \"ticket\", \"article\"",
                    "type": "string"
                },
                "rootId": {
                    "description": "Root comment ID for nested replies",
                    "type": "string"
                },
                "status": {
                    "description": "Moderation",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.CommentStatus"
                        }
                    ]
                },
                "tenantId": {
                    "description": "Multi-tenant support (e.g., \"shop\", \"ticket\", \"blog\")",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_minisource_comment_internal_models.CommentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected",
                "spam"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusApproved",
                "StatusRejected",
                "StatusSpam"
            ]
        },
...
        "github_com_minisource_comment_internal_models.EditRecord": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "editedAt": {
                    "type": "string"
                },
                "editedBy": {
                    "type": "string"
                }
            }
        },
...
        "github_com_minisource_comment_internal_models.ModerateCommentRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "rejectionReason": {
                    "type": "string"
                },
                "status": {
                    "enum": [
                        "approved",
                        "rejected",
                        "spam"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.CommentStatus"
                        }
                    ]
                }
            }
        },
        "github_com_minisource_comment_internal_models.PinCommentRequest": {
            "type": "object",
            "properties": {
                "isPinned": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_handler.BulkModerateRequest": {
            "type": "object",
            "properties": {
                "comment_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rejection_reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_minisource_comment_internal_models.CommentStatus"
                }
            }
        },
        "internal_handler.BulkModerateResponse": {
            "type": "object",
            "properties": {
                "failed_count": {
                    "type": "integer"
                },
                "failed_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "success_count": {
                    "type": "integer"
                }
            }
        },
        "internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}